update lua to 5.3.3

+++ app/src/main/java/org/keplerproject/luajava/LuaObject.java
@@ -40,14 +40,14 @@ import java.util.StringTokenizer;
  * </ul>
  * The LuaObject will represent only the object itself, not a variable or a stack index, so when you change a string,
  * remember that strings are immutable objects in Lua, and the LuaObject you have will represent the old one.
- * 
+ *
  * <h2>Proxies</h2>
- * 
+ *
  * LuaJava allows you to implement a class in Lua, like said before. If you want to create this proxy from Java, you
  * should have a LuaObject representing the table that has the functions that implement the interface. From this
  * LuaObject you can call the <code>createProxy(String implements)</code>. This method receives the string with the
  * name of the interfaces implemented by the object separated by comma.
- * 
+ *
  * @author Rizzato
  * @author Thiago Ponte
  */
@@ -59,7 +59,7 @@ public class LuaObject
 
 	/**
 	 * Creates a reference to an object in the variable globalName
-	 * 
+	 *
 	 * @param L
 	 * @param globalName
 	 */
@@ -76,7 +76,7 @@ public class LuaObject
 
 	/**
 	 * Creates a reference to an object inside another object
-	 * 
+	 *
 	 * @param parent
 	 *            The Lua Table or Userdata that contains the Field.
 	 * @param name
@@ -104,7 +104,7 @@ public class LuaObject
 
 	/**
 	 * This constructor creates a LuaObject from a table that is indexed by a number.
-	 * 
+	 *
 	 * @param parent
 	 *            The Lua Table or Userdata that contains the Field.
 	 * @param name
@@ -131,7 +131,7 @@ public class LuaObject
 
 	/**
 	 * This constructor creates a LuaObject from a table that is indexed by a LuaObject.
-	 * 
+	 *
 	 * @param parent
 	 *            The Lua Table or Userdata that contains the Field.
 	 * @param name
@@ -161,7 +161,7 @@ public class LuaObject
 
 	/**
 	 * Creates a reference to an object in the given index of the stack
-	 * 
+	 *
 	 * @param L
 	 * @param index
 	 *            of the object on the lua stack
@@ -186,7 +186,7 @@ public class LuaObject
 
 	/**
 	 * Creates the reference to the object in the registry table
-	 * 
+	 *
 	 * @param index
 	 *            of the object on the lua stack
 	 */
@@ -195,7 +195,7 @@ public class LuaObject
 		synchronized (L)
 		{
 			L.pushValue(index);
-			int key = L.Lref(LuaState.LUA_REGISTRYINDEX.intValue());
+			int key = L.Lref(LuaState.LUA_REGISTRYINDEX);
 			ref = new Integer(key);
 		}
 	}
@@ -207,7 +207,7 @@ public class LuaObject
 			synchronized (L)
 			{
 				if (L.getCPtrPeer() != 0)
-					L.LunRef(LuaState.LUA_REGISTRYINDEX.intValue(), ref.intValue());
+					L.LunRef(LuaState.LUA_REGISTRYINDEX, ref.intValue());
 			}
 		}
 		catch (Exception e)
@@ -221,7 +221,7 @@ public class LuaObject
 	 */
 	public void push()
 	{
-		L.rawGetI(LuaState.LUA_REGISTRYINDEX.intValue(), ref.intValue());
+		L.rawGetI(LuaState.LUA_REGISTRYINDEX, ref.intValue());
 	}
 
 	public boolean isNil()
@@ -389,7 +389,7 @@ public class LuaObject
 
 	/**
 	 * Calls the object represented by <code>this</code> using Lua function pcall.
-	 * 
+	 *
 	 * @param args -
 	 *            Call arguments
 	 * @param nres -
@@ -432,15 +432,15 @@ public class LuaObject
 				else
 					str = "";
 
-				if (err == LuaState.LUA_ERRRUN.intValue())
+				if (err == LuaState.LUA_ERRRUN)
 				{
 					str = "Runtime error. " + str;
 				}
-				else if (err == LuaState.LUA_ERRMEM.intValue())
+				else if (err == LuaState.LUA_ERRMEM)
 				{
 					str = "Memory allocation error. " + str;
 				}
-				else if (err == LuaState.LUA_ERRERR.intValue())
+				else if (err == LuaState.LUA_ERRERR)
 				{
 					str = "Error while running the error handler function. " + str;
 				}
@@ -452,7 +452,7 @@ public class LuaObject
 				throw new LuaException(str);
 			}
 
-			if (nres == LuaState.LUA_MULTRET.intValue())
+			if (nres == LuaState.LUA_MULTRET)
 				nres = L.getTop() - top;
 			if (L.getTop() - top < nres)
 			{
@@ -472,7 +472,7 @@ public class LuaObject
 
 	/**
 	 * Calls the object represented by <code>this</code> using Lua function pcall. Returns 1 object
-	 * 
+	 *
 	 * @param args -
 	 *            Call arguments
 	 * @return Object - Returned Object
@@ -519,7 +519,7 @@ public class LuaObject
 
 	/**
 	 * Function that creates a java proxy to the object represented by <code>this</code>
-	 * 
+	 *
 	 * @param implem
 	 *            Interfaces that are implemented, separated by <code>,</code>
 	 */
